+++ modules/generalStore/index.js
@@ -1,23 +1,42 @@
 'use strict'
-const redis = require('redis')
+const
+	redis = require('redis'),
+	spec = require('./index.spec')
 
 class GeneralStore {
-	constructor(settings) {
-		this.settings = settings
-
-		let options = {}
-		if (this.settings.redis.password) {
-			options.password = this.settings.redis.password
+	constructor(config) {
+		for (const testName in spec) {
+			this[testName] = spec[testName]
 		}
-		this.redisClient = redis.createClient(this.settings.redis.port, this.settings.redis.host, options)
+		this.config = config
+	}
+
+	createClient() {
+		return new Promise((resolve, reject) => {
+			let options = {}
+			if (this.config.redis.password) {
+				options.password = this.config.redis.password
+			}
+			this.client = redis.createClient(this.config.redis.port, this.config.redis.host, options)
+			this.client.on('ready', () => {
+				resolve(true)
+			})
+			this.client.on('error', (err) => {
+				reject(err)
+			})
+		})
 	}
 
 	getStoredEntry(handle) {
 		return new Promise((resolve, reject) => {
-			this.redisClient.hget('general_store', handle, (err, reply) => {
+			if ((typeof handle !== 'string') || !handle.length) {
+				reject({customMessage: 'Invalid handle provided.'})
+				return
+			}
+			this.client.hget('general_store', handle, (err, reply) => {
 				if (err) {
 					reject(err)
-					return;
+					return
 				}
 				resolve(reply)
 			})
@@ -26,10 +45,18 @@ class GeneralStore {
 
 	storeEntry(handle, entry) {
 		return new Promise((resolve, reject) => {
-			this.redisClient.hset('general_store', handle, entry, (err, reply) => {
+			if ((typeof handle !== 'string') || !handle.length) {
+				reject({customMessage: 'Invalid handle provided.'})
+				return
+			}
+			if (typeof entry === 'undefined') {
+				reject({customMessage: 'No entry value provided.'})
+				return
+			}
+			this.client.hset('general_store', handle, entry, (err, reply) => {
 				if (err) {
 					reject(err)
-					return;
+					return
 				}
 				resolve(reply)
 			})
@@ -38,10 +65,15 @@ class GeneralStore {
 
 	removeEntry(handle) {
 		return new Promise((resolve, reject) => {
-			this.redisClient.hdel('general_store', handle, (err, reply) => {
+			if ((typeof handle !== 'string') || !handle.length) {
+				reject({customMessage: 'Invalid handle provided.'})
+				return
+			}
+			this.client.hdel('general_store', handle, (err, reply) => {
 				if (err) {
+					console.log(e)
 					reject(err)
-					return;
+					return
 				}
 				resolve(reply)
 			})