+++ changelog.md
@@ -19,9 +19,19 @@
 	- *BREAKING* - removed the defaultConfig. From now on, ramster will blow up if you don't give it a config.
 - Added config templates.
 - Added templates for the following db modules: users.
-- Added a codeGenerator module, which generates config files, client server layout files and nginx config. Added tests for it.
+- Added a codeGenerator module, which generates config files, client server layout files and nginx config files. Added tests for it.
 - Added an .npmignore file, that keeps the "test" folder out of the final package (it's for testing ramster and you don't need it in the build). It's still in the repo, though.
 - Removed the buildLayoutFile and generateNGINXConfig methods from the client module, as the codeGenerator now covers this functionality.
+- *BREAKING* - Renamed cfg to config in db modules.
+- Added tests and validations to the generalStore and errorLogger modules.
+- A number of changes in the tokenManager, many of them *BREAKING*:
+	- *BREAKING* - signToken is now in the format signToken(userData, secret[, expiresInMinutes]), rather than signToken({userId, userData, secret, expiresInMinutes}).
+	- *BREAKING* - userId removed from signToken options; userData is now required, must be an object and must contain at least one key.
+	- *BREAKING* - verifyToken is now in the format verifyToken(token, secret), rather than verifyToken({token, secret}).
+	- *BREAKING* - createToken is now in the format createToken(type, userData, secret, moduleName[, expiresInMinutes]), rather than createToken({type, userId, userData, secret, moduleName, expiresInMinutes}).
+	- *BREAKING* - deleteToken is now in the format deleteTokens(userId, moduleName), rather than deleteTokens({userId, moduleName}).
+	- Added extra validations for the input parameters in all methods.
+	- Made some consistency and deprecation fixes.
 
 # 0.6.22
 - Updated the v1.0.0 roadmap with an additional feature.